Add mostly-barebones Connection and LogicComponent
LogicComponent is the abstract base class that contains Connections and
the evaluation logic. The evaluation logic will work like so: when
evaluate() is called, LogicComponent will look at all its input
Connections and find the appropriate value using the same evaluate()
method. It will then call an abstract logicalEvaluate() method to
actually perform the required logic. (Input/Output will override both.)
The result will then be passed on. An EvaluationListener will eventually
be used so that UI updates on evaluation may be shown.

Connection will represent a connection between a specific output on one
LogicComponent and a specific input on another. It is currently a
skeleton class to make LogicComponent compile.