Add useful methods in Connection
Connection now has getters for the input and output ports and implements
equals() and hashCode(). It also has a getValue() method which calls
the output port's LogicComponent's evaluate() method and selects the
appropriate output value. It will be LogicComponent.evaluate()'s
responsibility to memoize its results for each input combination.