Give LogicComponents UUIDs
LogicComponents needed to be identified by something other than their
object's position in memory. Each LogicComponent now has a UUID and
implement equals() and hashCode() based on that UUID. toString() now
uses the UUID and the abstract toString() has been replaced with an
abstract getName() method which is then called by LogicComponent's
toString() and added to the UUID to produce the final toString() result.