Add Gate class between gates and LogicComponent
Gate is a subclass of LogicComponent and the parent of the various gate
classes (AndGate, NotGate, OrGate, NandGate). The complex evaluate()
method which assembles input values into an array was moved to Gate;
memoization will also be done only in Gate and not in LogicComponent.
This is because Input and Output, the other two LogicComponent
subclasses, have no need of the complex evaluate() method or
memoization.

LogicComponent.evaluate() is now abstract, and
LogicComponent.logicalEval() does not exist. logicalEval() is now an
abstract method in Gate.