Add memoization for Gates
Gates now cache their last result in the memoizedOutput member variable.
This cache may be cleared with the LogicComponent.markDirty() method
which is only implemented in Gate. This is used by Input. When
Input.setValue() is called, the output of any gates connected to the
Input throughout the circuit graph is no longer known, so Input calls
markDirty() recursively traversing through the graph. It ends its
recursion either when encountering an Output, a null Port.Output, or an
already-dirty node, which Input assumes to have been dirtied by another
Input.setValue() call already and does not change. This should improve
performance in large circuits with many Outputs.