Add EvaluationListener with tests
EvaluationListener is a functional interface with one method,
onEvaluate(), with one parameter, a EvaluationListener.Event. This
inner class is an immutable POJO/data object with a LogicComponent where
the evaluation took place as well as the relevant input and output
arrays.

A possibly-null EvaluationListener is passed to
LogicComponent.evaluate() and if non-null is called upon the successful
completion of an evaluation. This means that onEvaluate() will be called
from bottom to top: the lowest non-memoized component in the gate will
be the first to have onEvaluate() called as it will be the first to
finish evaluating. Inputs, Outputs, and Gates all call the listener. The
listener is passed through Port.Input.getInputValue() and
Connection.getValue() and will not be called for memoized components.