Add Position and basic FieldComponent
Position is an immutable POJO that represents a position on the field.
It has x and y double values which may be queried but not updated, and
it implements equals()/hashCode()/toString().

FieldComponent will represent something actually on the field. It
contains a LogicComponent and a changeable Position. It's currently just
a data object but will have behaviour (such as saving, types of input/
output, connecting to other FieldComponents, etc.) in the future.