Add InputFC
InputFC (standing for InputFieldComponent) is a subclass of
FieldComponent encapsulating an Input. It has a key binding (which is
initially negative, representing no binding, until set) and a type
(defined in an inner enum InputFC.Type, either BUTTON or SWITCH). It
also contains a toggle() method for toggling the underlying Input. In
the future it will also contain references to graphics for each type of
InputFC.

Field was also updated with a method getInputFCs() which returns all
FieldComponents in its list that are InputFCs.