Add OutputFC
OutputFC represents an output on the field. It encapsulates the logic
layer's Output and has an OutputFC.Type (an enum), which tells what kind
of output it is: a lamp, a binary 1/0, or a red/green/blue pixel.

OutputFC contains a package-private evaluate(EvaluationListener) method
that simply delegates to the underlying Output, returning a boolean
instead of a boolean[1] for convenience. It should only be called by
Field in the process of evaluation.