Add Field.getOutputFCs() and FieldTest
Field.getOutputFCs() simply returns all FieldComponents on the Field
that are OutputFCs. It performs the same function as Field.getInputFCs()
but for OutputFC. getInputFCs() and getOutputFCs() are combined into the
generic private method getFCsOfType() which uses generics and a passed
Class parameter to get either InputFCs or OutputFCs.

FieldTest contains several tests of getInputFCs() and getOutputFCs() to
ensure that getFCsOfType()'s generics work properly.