Add UserGate
UserGate is a gate customized by the user. It contains a Field to which
it delegates evaluation. UserGates will be creatable by the user based
on a Field. They will also be savable.

As well, Field now maintains Lists of its InputFCs and OutputFCs.
getInputFCs() and getOutputFCs() now return unmodifiable versions of
these lists instead of delegating to a common reflection-based
generic method. This was done so that Field has an order to its
InputFCs and OutputFCs so that UserGate can determine which input to
set to which InputFCs. Field now also has an evaluate() method which
simply returns an array of the results of the evaluation of each
OutputFC in order.