Add split-screen layout with tool and field panes
The UI is now split into two sections, ToolPane and FieldPane, using
SplitPane. ToolPane will contain a list of radio-button-behaviour
image-button tools that the user can select to perform certain actions.
FieldPane will contain the actual field.

Also, Tool has been added. This is an abstract base class for a tool,
which is something that performs an action on the field. Tools are added
to the ToolPane at runtime based on the contents of Tool.ALL_TOOLS, a
constant array. Tools have various hooks for when they are selected and
deselected, as well as when the user hovers over or clicks on the field
with the tool selected.