Add PlaceComponentTool, UIComponent
UIComponent is a subclass of Group and is a UI representation of a
FieldComponent. It consists (or at least is supposed to consist) of a
square surrounding the name of the UIComponent with circles on the left
and right sides to represent the input and output ports respectively. It
has a FieldComponent.

A UIComponent may be a "ghost". If it is, it is a lighter grey and is
transparent to mouse events. Ghost UIComponents are used to show a
preview of where a UIComponent may be placed when the user hovers over
the FieldPane.

Tool has two abstract methods, onHover() and onClick(). They are called
when the tool is selected and the user hovers or clicks on the FieldPane
(which is now a Pane rather than an AnchorPane). This is done by
ToolPaneController.

PlaceComponentTool is an implementation of Tool which places a
FieldComponent. It has a "ghost" UIComponent which it moves as the user
hovers over the FieldPane. When the user clicks, it creates a
FieldComponent, adds it to the Field in the main LogikosApplication
instance, creates a UIComponent using the FieldComponent, and adds the
UIComponent to the FieldPane.

As well, LogikosApplication now contains the selected tool and the
global Field. It injects itself into both FieldPaneController and
ToolPaneController, which are subclasses of the new class
MainPaneController which facilitates this. The LogikosApplication
instance is passed to Tools as well.

Also, a null check was removed from FieldComponent.setPosition() as the
ghost UIComponents have null position.