Fix ghost staying on FieldPane after mouse leaves
Previously the ghost UIComponent would stay on the FieldPane after the
mouse left it, and would stay on after another tool was selected. This
was fixed by adding an onLeavePane() hook in Tool called by another
onLeavePane() hook in FieldPaneController. Also, onSelect() and
onDeselect() now require a LogikosApplication instance as the second
part of this bug was fixed by removing the ghost UIComponent in
PlaceComponentTool.onDeselect().