Add PannablePane for FieldPane
The FieldPane is now a PannablePane, a custom Pane which allows the user
to pan with the right mouse button.

PannablePane works by putting all its content inside a wrapped Pane
called "content". Nodes must be added to a PannablePane through
getContentChildren() rather than getChildren(). As well, to convert a
pane-relative coordinate on the screen to an "absolute" coordinate
for placing a node on the PannablePane, use paneToRealX() and
paneToRealY().