Fix component detection in click/hover handling
Previously, the attempt to detect what FieldComponent the user's mouse
was hovering over or clicking on in FieldPaneController's
callSelectedToolHook() method failed miserably. Now, event filters are
used in UIComponent to fix this bug, and the hovered/clicked component
is now detected reliably.

As well, the relevant parameter has been changed from a FieldComponent
to a UIComponent as it makes little sense to find the FieldComponent
the user is hovering over/clicking on.