Make it impossible for components to overlap
PlaceComponentTool now has a private method, doesGhostOverlapAnything(),
which iterates through the UIComponent nodes of the FieldPane and
returns whether the ghost UIComponent overlaps any of them. It then sets
the ghost to invisible if it overlaps anything and will not place a
UIComponent when the ghost overlaps something. Thus, UIComponents may no
longer overlap each other.