Add a simple SelectTool
SelectTool is a Tool which allows the user to select a UIComponent and
interact with it. When the user clicks on a UIComponent with SelectTool
selected, the UIComponent enters a "selected" visual state and is stored
in the SelectTool's selectedComponent member variable. The only
interaction currently implemented is deletion: when the user presses the
delete key with a UIComponent selected, the component is removed from
the FieldPane and its FieldComponent is removed from the Field.

As well, UIComponent#setSelected() and #setDeselected() were added.
These simply add and remove a drop shadow, respectively, and are used to
indicate when a UIComponent is in a selected state in SelectTool.