Add MousePosition to encapsulate mouse data
MousePosition is a POJO used to encapsulate the data about the mouse's
position on the FieldPane at the time of an event. A MousePosition
contains the mouse's x and y coordinates relative to the FieldPane
(paneX and paneY), as well as the UIComponent it was over (or null;
component) and an instance of PortOver, an inner class (or null)
containing data about the port the mouse was over.

MousePosition.PortOver is also a POJO which contains data about the port
the user's mouse was over at the time of the event, if it was over a
port (otherwise PortOver variables should be null). It encapsulates the
port's type (input or output) and number.

This is useful so as not to continue adding parameters onto onClick()
and onHover() in Tool and to pave the way for port-based selection for
a future ConnectTool.