Add connections on the UI level
There are now connections in the UI layer.

UIConnection is a Node which visually represents a connection. It has
two UIComponents ("from" and "to") as well as the port numbers of the
corresponding input/output ports. It draws small circles inside each
port node on the UIComponents and connects them with a line.

ConnectTool is a Tool for establishing a UIConnection between two
UIComponents. The user clicks on a port of one UIComponent then the
opposite kind of port on the other UIComponent. ConnectTool creates a
UIConnection between the two and also establishes a Connection between
the corresponding LogicComponents in the logic layer.

Care had to be taken in the case that the user connects to an input port
where a connection already exists. Since input ports cannot have
multiple connections, the previous connection had to be removed at both
the UI and logic layers. UIConnection.destruct() was created and
Connection.destruct() and InputPort.removeConnection() were added to
facilitate this.

Also, in UIComponent, the lists of input/output ports are now arrays and
there are arrays of input/output connections. The input connection array
is simply a UIConnection[], but the output connection array (since
outputs may have multiple connections) is a List<UIConnection>[] that
is initialized with empty ArrayLists in the constructor.

Note that connections are not yet deleted when a UIComponent is deleted
with SelectTool, and they cannot yet be deleted on their own.