Add Selectable, make UIConnection selectable
Selectable is an interface which represents a UI node which can be
selected with the SelectTool. It is used inside SelectTool for the
selected item.

UIComponent and UIConnection implement Selectable, and SelectTool is now
called and MouseEvents are created for those on UIConnections.
UIConnections can thereby be selected and deleted.