Add input components on UI level
There is now a UIComponent subclass, InputUIC, representing an input on
the UI level. It maps InputFC.Types to images (all added under
src/main/resources) and shows the image as the component.

PlaceInputTool is a PlaceComponentTool subclass whose only job is
currently to place InputUICs instead of plain UIComponents.
PlaceComponentTool has been generalized with function references to
yield a minimum amount of code in PlaceInputTool.