Allow PlaceInputTool to scroll through input types
A new global event hook has been added to Tool, onScroll(). It takes a
ScrollEvent and like onKeyPress() is called globally.

PlaceInputTool uses this hook to scroll through the possible input types
on scrolling. It does this by strategically replacing the fcMaker (now
accessible by a protected method in PlaceComponentTool) and then calling
the new makeGhost() method, which regenerates the ghost component with
the new fcMaker.

As well, the ghost InputUIC is now only 70% greyscale, making the
differently coloured buttons distinguishable.