Add connection ghost in ConnectTool
There is now a 'ghost' UIConnection used in ConnectTool to provide a
visual indicator of connecting. The ghost UIConnection is created with
UIConnection.ghost() and has very little functionality.

To facilitate the reassignment of the ghost UIConnection's toCenterX/Y
in onHover() in ConnectTool so that it moves with the mouse,
UIConnection now has properties for this bound to the relevant nodes'
properties. These are toCenterX/YProperty, accessible via various
methods in UIConnection.