Prevent circular connections
ConnectTool now checks to make sure connections are not circular before
establishing a connection. It does this by traversing the tree forwards
from the "input" end of the to-be-established connection, and if the
"output" end is reached, the connection is declared invalid and is
stopped. This prevents both ridiculous scenarios like connecting a gate
to itself and more involved circular connections.