Add evaluation box, which can contain Tools
The EvaluationBox is a box in the top-right of the FieldPane which
contains evaluation tools. Its controller is EvaluationBoxController.

Tools now also have "location" and "enabled" attributes. Tool.Location
is an enum which describes where the tool is to be placed: the ToolPane
or the EvaluationBox. Each tool has a location. The enabled property
describes whether it is enabled, and is bound to the various radio
buttons' disableProperty (inverted).

As well, since the radio button creation logic would be duplicated
between ToolPaneController and EvaluationBoxController, the relevant
logic has been moved to Tool.createRadioButton().

StartEvaluateTool is an as-of-now skeleton tool placed in the
EvaluationBox.