Add UI evaluation logic
StartEvaluationTool has been renamed to RunTool and now actually
performs some evaluation when active.

RunTool calls Field.evaluate() and matches the results with the
OutputUICs, on which it then calls the new method
BooleanStateImageUIC.setState() to change the image.

Note that changing the inputs from off or colouring the connections is
not yet supported.