Add support for pressable InputFCs in RunTool
RunTool now supports pressable InputFC types (i.e. those with behaviour
InputFC.Type.Behaviour.PRESS). This is done through two new hooks in
Tool, onPress() and onRelease(), implemented largely the same as the
other mouse hooks.

When RunTool.onPress() is called, it sets the relevant InputUIC to "on"
and stores the InputUIC in a member variable. When RunTool.onRelease()
is then called, it sets the InputUIC in that member variable to "off"
and clears the variable. In this way RunTool is able to release the
correct InputUIC irregardless of the position of the mouse when
onRelease() is called.